refactor(module-http): flatten mock/adapters into mock/, fix lint warnings
mock/ only ever had one barrel (index.ts) plus an adapters/ subfolder holding
the 3 files it re-exported -- two barrels doing the same job for exactly one
category of file. Moved create-open-api-mock-middleware.ts,
create-router-middleware.ts and resolve-open-api-mock-response.ts up into
mock/ directly and dropped the adapters/ barrel, fixing relative import
depths and OpenApiMockLike's re-export path along the way.

Also fixes the fusion-lint warnings the files carried before the move:
renamed each file to match its exported symbol (filename-convention),
added the missing intent comments on create-router-middleware.ts's
if-blocks and .map() (require-intent-comment), and replaced a comma-operator
expression in compilePath with an explicit if/return (noCommaOperator).
